refactor: query key를 별도의 파일로 분리하여 관리 (#ATR-400)
