--- apps/service/features/trend-newsletters/lib/index.ts
@@ -1,1 +1,0 @@
-export * from './hooks'
