+++ apps/service/features/trend-newsletters/model/hooks/index.ts
