+++ apps/service/features/trend-newsletters/model/hooks/useTrendNewsletters.ts
@@ -1,13 +1,14 @@
 import { UseSuspenseQueryResult, useSuspenseQuery } from '@tanstack/react-query'
 import { fetchNewsletters } from '../../api'
-import { MainCategory, TrendNewsletterResponse } from '../../model'
+import { MainCategory, TrendNewsletterResponse } from '../type'
+import trendNewsletterQueryKeys from '../trendNewsletterQueryKeys'
 
 export default function useTrendNewsletters(
   category: MainCategory,
   size: number = 10,
 ): UseSuspenseQueryResult<TrendNewsletterResponse, Error> {
   return useSuspenseQuery({
-    queryKey: ['trend-newsletters', category, size],
+    queryKey: trendNewsletterQueryKeys.trendNewsletters({ category, size }),
     queryFn: () => fetchNewsletters(category, size),
   })
 }
