+++ apps/service/features/trend-newsletters/model/index.ts
@@ -1,1 +1,2 @@
 export * from './type'
+export * from './hooks'
