+++ apps/service/features/trend-newsletters/model/trendNewsletterQueryKeys.ts
@@ -1,0 +1,14 @@
+import { TrendNewsletterParams } from './type'
+
+const trendNewsletterQueryKeys = {
+  trendNewsletters: (params: TrendNewsletterParams) => [
+    'trend-newsletters',
+    params,
+  ],
+  preferCategories: (params: string | undefined) => [
+    'prefer-categories',
+    params,
+  ],
+}
+
+export default trendNewsletterQueryKeys
