+++ apps/service/widgets/main/model/mainQueryKeys.ts
@@ -1,0 +1,5 @@
+const mainQueryKeys = {
+  recentArticles: (params: string | undefined) => ['recent-articles', params],
+}
+
+export default mainQueryKeys
